+++ ChangeLog.md
@@ -1,3 +1,1 @@
-- The city wide traffic percentage is now calculated using road importance, higher laned roads have a bigger effect on the percentage. Which should improve the abnormally high traffic percentages shown before and show a more accurate reflection.
-
-- Live traffic flow now feeds the vanilla traffic flow graph (Will need a full 24 ingame hour simulation to populate the graph).
+- Added a 24 hour average percentage from the live congestion percentage data.